+++ frontend/src/components/SignupFormPage/SignupForm.css
@@ -1,4 +1,32 @@
 .signUpForm{
     display: flex;
     flex-direction: column;
+    justify-content: center;
+    font-family: Arial, Helvetica, sans-serif;
+    align-items: center;
+    gap: 10px;
+}
+
+.signupInput{
+    width: 250px;
+    height: 25px;
+    font-size: large;
+    border-radius: 5px;
+    border: 1px solid rgb(119, 121, 121);
+}
+
+.signupButton{
+    color: white;
+    background-color: rgb(113, 62, 172);
+    border: none;
+    border-radius: 5px;
+    width: 150px;
+    font-size: medium;
+    padding: 5px;
+    margin-top: 10px;
+}
+
+.signupHeader{
+    margin-bottom: 15px;
+    margin-top: 0;
 }
