+++ frontend/src/context/Modal.css
@@ -22,7 +22,8 @@
     position: absolute;
     background-color:white;
     width: 300px;
-    height: 350px;
+    height: fit-content;
+    min-height: 350px;
     padding:10px;
     border-radius: 5px;
     display: flex;
